+++ src/asm/functions/sub_asm.c
@@ -11,6 +11,6 @@ bool sub_asm(__attribute__((unused)) int fd)
 {
 	uint8_t i = 5;
 
-	write(fd, &i, sizeof(int));
+	write(fd, &i, sizeof(uint8_t));
 	return true;
 }
